feat(client): update RecentlyPlayed component and create hook
